Tidy up and update snapshots
