+++ samples/WebApplication/WebApplication.csproj
@@ -19,7 +19,7 @@
     </PropertyGroup>
 
     <ItemGroup>
-        <PackageReference Include="Refit" Version="7.0.0"/>
+        <PackageReference Include="Refit" Version="8.0.0"/>
         <PackageReference Include="System.Text.Json" Version="8.0.5" />
     </ItemGroup>
