+++ samples/WebApplicationConsumer/WebApplicationConsumer.csproj
@@ -11,7 +11,7 @@
     <ItemGroup>
         <PackageReference Include="System.Text.Json" Version="8.0.5" />
         <PackageReference Include="Newtonsoft.Json" Version="13.0.3" />
-        <PackageReference Include="Refit" Version="7.0.0" />
+        <PackageReference Include="Refit" Version="8.0.0" />
         <PackageReference Include="ServiceStack.Text" Version="8.2.2" />
     </ItemGroup>
