+++ src/Vogen/ValueObjectGenerator.cs
@@ -128,7 +128,7 @@ public class ValueObjectGenerator : IIncrementalGenerator
         
         // the user can specify to create the MessagePack generated code as an attribute
         // or as marker in another project.
-        GenerateCodeForMessagePack.GenerateForApplicableValueObjects(spc, compilation, workItems, vogenKnownSymbols);
+        GenerateCodeForMessagePack.GenerateForApplicableValueObjects(spc, compilation, workItems);
         GenerateCodeForMessagePackMarkers.GenerateForMarkerClasses(spc, compilation, markerClasses, vogenKnownSymbols);
         
         GenerateCodeForBsonSerializers.WriteIfNeeded(spc, compilation, workItems);
