+++ tests/ConsumerTests/BugFix624.cs
@@ -1,8 +1,3 @@
-using System.Text.Json;
-using System.Text.Json.Serialization;
-using System.Text.Json.Serialization.Metadata;
-using SystemTextJsonSerializer = System.Text.Json.JsonSerializer;
-
 namespace ConsumerTests.BugFixTests.BugFix639;
 
 [ValueObject<string>(fromPrimitiveCasting: CastOperator.Implicit)]
