+++ tests/ConsumerTests/BugFixTests/BugFix624.cs
@@ -1,7 +1,4 @@
 using System.Text.Json;
-using System.Text.Json.Serialization;
-using System.Text.Json.Serialization.Metadata;
-using SystemTextJsonSerializer = System.Text.Json.JsonSerializer;
 
 namespace ConsumerTests.BugFixTests.BugFix624;
