+++ tests/ConsumerTests/SerializationAndConversionTests/ComplexSerializationTests_SytemTextJson.cs
@@ -6,7 +6,7 @@ using Vogen.IntegrationTests.TestTypes;
 
 namespace MediumTests.SerializationAndConversionTests;
 
-public partial class ComplexSerializationTests
+public partial class ComplexSerializationTests_ServiceStackDotText
 {
     public class ComplexSystemTextJson
     {
